pip-requirements.txt
certifi==2019.11.28
cycler==0.10.0
Cython==0.29.15
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==6.2.2
pycocotools==2.0
pyparsing==2.4.6
python-dateutil==2.8.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torchsummary==1.5.1
tqdm==4.43.0
wincertstore==0.2
scikit-image==0.16.2
